+++ madtranslate/__init__.py
@@ -3,11 +3,11 @@ from pathlib import Path
 
 from redbot.core.bot import Red
 
-from .stattrack import StatTrack
+from .madtranslate import MadTranslate
 
-with open(Path(__file__).parent / "info.json", encoding="utf8") as fp:
+with open(Path(__file__).parent / "info.json") as fp:
     __red_end_user_data_statement__ = json.load(fp)["end_user_data_statement"]
 
 
 def setup(bot: Red) -> None:
-    bot.add_cog(StatTrack(bot))
+    bot.add_cog(MadTranslate(bot))