+++ system/system.py
@@ -1,9 +1,11 @@
 import datetime
+from typing import List
 
 import discord
 import psutil
 from redbot.core import commands
 from redbot.core.bot import Red
+from redbot.core.utils.chat_formatting import humanize_timedelta
 from vexcogutils import format_help, format_info
 
 from .command import DynamicHelp
@@ -17,6 +19,7 @@ from .utils import (
     get_sensors,
     get_uptime,
     get_users,
+    up_for,
 )
 
 UNAVAILABLE = "\N{CROSS MARK} This command isn't available on your system."
@@ -33,7 +36,7 @@ class System(commands.Cog):
     See the help for individual commands for detailed limitations.
     """
 
-    __version__ = "1.2.1"
+    __version__ = "1.2.4"
     __author__ = "Vexed#3211"
 
     def __init__(self, bot: Red) -> None:
@@ -50,6 +53,45 @@ class System(commands.Cog):
     @commands.command(hidden=True)
     async def systeminfo(self, ctx: commands.Context):
         await ctx.send(await format_info(self.qualified_name, self.__version__))
+
+    def finalise_embed(self, e: discord.Embed) -> discord.Embed:
+        """Make embeds look nicer - limit to two columns and set the footer to boot time"""
+        # needed because otherwise they are otherwise too squashed together so tabulate breaks
+        # doesn't look great on mobile but is fully bearable, more than ugly text wrapping
+
+        # oh, don't mention the ugly code please :P
+        # it works...
+        emb = e.to_dict()
+        fields: List[dict] = emb["fields"]
+        if len(fields) > 2:  # needs multi rows
+            data: List[List[dict]] = []
+            temp = []
+            for field in fields:
+                temp.append(field)
+                if len(temp) == 2:
+                    data.append(temp)
+                    temp = []
+            if len(temp) != 0:  # clear up stragglers
+                data.append(temp)
+
+            empty_field = {"inline": True, "name": ZERO_WIDTH, "value": ZERO_WIDTH}
+            fields = []
+            row: List[dict]
+            for row in data:
+                while len(row) < 3:
+                    row.append(empty_field)
+                fields.extend(row)
+
+        # else it's 2 or less columns so doesn't need special treatment
+        emb["fields"] = fields
+        e = discord.Embed.from_dict(emb)
+
+        # and footer is just a nice touch, thanks max for the idea of uptime there
+        sys_uptime = humanize_timedelta(seconds=up_for())
+        bot_uptime = humanize_timedelta(timedelta=datetime.datetime.now() - self.bot.uptime)
+        e.set_footer(text=f"System uptime: {sys_uptime}\nBot uptime: {bot_uptime}")
+
+        return e
 
     @commands.is_owner()
     @commands.group()
@@ -80,15 +122,12 @@ class System(commands.Cog):
             time = data["time"]
             freq = data["freq"]
             if await ctx.embed_requested():
-                now = datetime.datetime.utcnow()
-                embed = discord.Embed(
-                    title="CPU Metrics", colour=await ctx.embed_colour(), timestamp=now
-                )
+                embed = discord.Embed(title="CPU Metrics", colour=await ctx.embed_colour())
                 embed.add_field(name="CPU Usage", value=box(percent))
                 embed.add_field(name="CPU Times", value=box(time))
                 extra = data["freq_note"]
-                embed.add_field(name=f"CPU Frequency{extra}", value=box(freq), inline=False)
-                await ctx.send(embed=embed)
+                embed.add_field(name=f"CPU Frequency{extra}", value=box(freq))
+                await ctx.send(embed=self.finalise_embed(embed))
             else:
                 msg = "**CPU Metrics**\n"
                 to_box = f"CPU Usage\n{percent}\n"
@@ -99,8 +138,8 @@ class System(commands.Cog):
                 await ctx.send(msg)
 
     @system.command(
-        name="mem", aliases=["memory", "ram"], cls=DynamicHelp, supported_sys=True
-    )  # all systems
+        name="mem", aliases=["memory", "ram"], cls=DynamicHelp, supported_sys=True  # all systems
+    )
     async def system_mem(self, ctx: commands.Context):
         """
         Get infomation about memory usage.
@@ -114,11 +153,10 @@ class System(commands.Cog):
         physical = data["physical"]
         swap = data["swap"]
         if await ctx.embed_requested():
-            now = datetime.datetime.utcnow()
-            embed = discord.Embed(title="Memory", colour=await ctx.embed_colour(), timestamp=now)
+            embed = discord.Embed(title="Memory", colour=await ctx.embed_colour())
             embed.add_field(name="Physical Memory", value=box(physical))
             embed.add_field(name="SWAP Memory", value=box(swap))
-            await ctx.send(embed=embed)
+            await ctx.send(embed=self.finalise_embed(embed))
         else:
             msg = "**Memory**\n"
             to_box = f"Physical Memory\n{physical}\n"
@@ -149,11 +187,10 @@ class System(commands.Cog):
         temp = data["temp"]
         fans = data["fans"]
         if await ctx.embed_requested():
-            now = datetime.datetime.utcnow()
-            embed = discord.Embed(title="Sensors", colour=await ctx.embed_colour(), timestamp=now)
+            embed = discord.Embed(title="Sensors", colour=await ctx.embed_colour())
             embed.add_field(name="Temperatures", value=box(temp))
             embed.add_field(name="Fans", value=box(fans))
-            await ctx.send(embed=embed)
+            await ctx.send(embed=self.finalise_embed(embed))
         else:
             msg = "**Temperature**\n"
             to_box = f"Temperatures\n{temp}\n"
@@ -177,11 +214,10 @@ class System(commands.Cog):
         if not data:
             return await ctx.send("It looks like no one is logged in.")
         if embed:
-            now = datetime.datetime.utcnow()
-            embed = discord.Embed(title="Users", colour=await ctx.embed_colour(), timestamp=now)
+            embed = discord.Embed(title="Users", colour=await ctx.embed_colour())
             for name, userdata in data.items():
                 embed.add_field(name=name, value=box(userdata))
-            await ctx.send(embed=embed)
+            await ctx.send(embed=self.finalise_embed(embed))
         else:
             msg = "**Users**\n"
             to_box = "".join(f"{name}\n{userdata}" for name, userdata in data.items())
@@ -189,8 +225,8 @@ class System(commands.Cog):
             await ctx.send(msg)
 
     @system.command(
-        name="disk", aliases=["df"], cls=DynamicHelp, supported_sys=True
-    )  # all systems
+        name="disk", aliases=["df"], cls=DynamicHelp, supported_sys=True  # all systems
+    )
     async def system_disk(self, ctx: commands.Context):
         """
         Get infomation about disks connected to the system.
@@ -207,11 +243,10 @@ class System(commands.Cog):
         data = await get_disk(embed)
 
         if embed:
-            now = datetime.datetime.utcnow()
-            embed = discord.Embed(title="Disks", colour=await ctx.embed_colour(), timestamp=now)
+            embed = discord.Embed(title="Disks", colour=await ctx.embed_colour())
             for name, diskdata in data.items():
                 embed.add_field(name=name, value=box(diskdata))
-            await ctx.send(embed=embed)
+            await ctx.send(embed=self.finalise_embed(embed))
         else:
             msg = "**Disks**\n"
             to_box = "".join(f"{name}\n{diskdata}" for name, diskdata in data.items())
@@ -219,8 +254,8 @@ class System(commands.Cog):
             await ctx.send(msg)
 
     @system.command(
-        name="processes", aliases=["proc"], cls=DynamicHelp, supported_sys=True
-    )  # all systems
+        name="processes", aliases=["proc"], cls=DynamicHelp, supported_sys=True  # all systems
+    )
     async def system_processes(self, ctx: commands.Context):
         """
         Get an overview of the status of currently running processes.
@@ -231,20 +266,17 @@ class System(commands.Cog):
             proc = (await get_proc())["statuses"]
 
         if await ctx.embed_requested():
-            now = datetime.datetime.utcnow()
-            embed = discord.Embed(
-                title="Processes", colour=await ctx.embed_colour(), timestamp=now
-            )
+            embed = discord.Embed(title="Processes", colour=await ctx.embed_colour())
             embed.add_field(name="Status", value=box(proc))
-            await ctx.send(embed=embed)
+            await ctx.send(embed=self.finalise_embed(embed))
         else:
             msg = "**Processes**\n"
             msg += box(f"CPU\n{proc}\n")
             await ctx.send(msg)
 
     @system.command(
-        name="network", aliases=["net"], cls=DynamicHelp, supported_sys=True
-    )  # all systems
+        name="network", aliases=["net"], cls=DynamicHelp, supported_sys=True  # all systems
+    )
     async def system_net(self, ctx: commands.Context):
         """
         Get network stats. They may have overflowed and reset at some point.
@@ -254,18 +286,17 @@ class System(commands.Cog):
         stats = (await get_net())["counters"]
 
         if await ctx.embed_requested():
-            now = datetime.datetime.utcnow()
-            embed = discord.Embed(title="Network", colour=await ctx.embed_colour(), timestamp=now)
+            embed = discord.Embed(title="Network", colour=await ctx.embed_colour())
             embed.add_field(name="Network Stats", value=box(stats))
-            await ctx.send(embed=embed)
+            await ctx.send(embed=self.finalise_embed(embed))
         else:
             msg = "**Network**\n"
             msg += box(f"Network Stats\n{stats}\n")
             await ctx.send(msg)
 
     @system.command(
-        name="uptime", aliases=["up"], cls=DynamicHelp, supported_sys=True
-    )  # all systems
+        name="uptime", aliases=["up"], cls=DynamicHelp, supported_sys=True  # all systems
+    )
     async def system_uptime(self, ctx: commands.Context):
         """
         Get the system boot time and how long ago it was.
@@ -275,18 +306,17 @@ class System(commands.Cog):
         uptime = (await get_uptime())["uptime"]
 
         if await ctx.embed_requested():
-            now = datetime.datetime.utcnow()
-            embed = discord.Embed(title="Uptime", colour=await ctx.embed_colour(), timestamp=now)
+            embed = discord.Embed(title="Uptime", colour=await ctx.embed_colour())
             embed.add_field(name="Uptime", value=box(uptime))
-            await ctx.send(embed=embed)
+            await ctx.send(embed=self.finalise_embed(embed))
         else:
             msg = "**Utime**\n"
             msg += box(f"Uptime\n{uptime}\n")
             await ctx.send(msg)
 
     @system.command(
-        name="top", aliases=["overview", "all"], cls=DynamicHelp, supported_sys=True
-    )  # all systems
+        name="top", aliases=["overview", "all"], cls=DynamicHelp, supported_sys=True  # all systems
+    )
     async def system_all(self, ctx: commands.Context):
         """
         Get an overview of the current system metrics, similar to `top`.
@@ -309,16 +339,13 @@ class System(commands.Cog):
             procs = proc["statuses"]
 
         if await ctx.embed_requested():
-            now = datetime.datetime.utcnow()
-            embed = discord.Embed(title="Overview", colour=await ctx.embed_colour(), timestamp=now)
+            embed = discord.Embed(title="Overview", colour=await ctx.embed_colour())
             embed.add_field(name="CPU Usage", value=box(percent))
             embed.add_field(name="CPU Times", value=box(times))
-            embed.add_field(name=ZERO_WIDTH, value=ZERO_WIDTH)
             embed.add_field(name="Physical Memory", value=box(physical))
             embed.add_field(name="SWAP Memory", value=box(swap))
-            embed.add_field(name=ZERO_WIDTH, value=ZERO_WIDTH)
             embed.add_field(name="Processes", value=box(procs))
-            await ctx.send(embed=embed)
+            await ctx.send(embed=self.finalise_embed(embed))
         else:
             msg = "**Overview**\n"
             to_box = f"CPU\n{cpu}\n\n"